Fixed compiler warnings left in cluster & raft modules

* 'fix/cluster-warnings-v1' of github.com:mmaslankaprv/v:
  r/vote_stm: fixed warnings
  c/partition_manager: removed unused shard_table field
  cluster: fixed warnings in cluster module

+++ src/v/cluster/partition_allocator.cc
@@ -58,8 +58,8 @@ partition_allocator::allocate_replicas(int16_t replication_factor) {
     std::vector<model::broker_shard> replicas;
     replicas.reserve(replication_factor);
 
-    while (replicas.size() < replication_factor) {
-        const int16_t replicas_left = replication_factor - replicas.size();
+    while (replicas.size() < (size_t)replication_factor) {
+        const uint16_t replicas_left = replication_factor - replicas.size();
         if (_available_machines.size() < replicas_left) {
             rollback(replicas);
             return std::nullopt;
